Moved some methods out of phyx2ontology.js into a separate file.
